requirements.txt
asgiref==3.2.5
confusable-homoglyphs==3.2.0
Django==3.0.4
django-bootstrap3==12.0.3
django-registration==3.1
django-tinymce==2.8.0
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
